requirements.txt
requests
pydub
google-cloud-speech
google-cloud-texttospeech